@@ -9,17 +9,16 @@ use Misery\Component\Connections\E5Dal\Client\E5DalAPIAccount;
 
 class ApiClientFactory implements RegisteredByNameInterface
 {
-    private ?MemoryContext $memoryContext = null;
-
     public function createFromConfiguration(array $account): ApiClientInterface
     {
         $type = $account['type'] ?? null;
+        $domain =  rtrim($account['domain'], '/');
         if ($type === 'basic_auth') {
             try {
                 // no need to authorize a basic auth
 
                 return new BasicAuthApiClient(
-                    $account['domain'],
+                    $domain,
                     $account['username'],
                     $account['password']
                 );
@@ -30,7 +29,7 @@ class ApiClientFactory implements RegisteredByNameInterface
         if ($type === 'microsoft_oauth') {
             try {
                 // no need to authorize a basic auth
-                $client = new ApiClient($account['domain']);
+                $client = new ApiCurlClient($domain);
 
                 $account = new MicrosoftDynamicsOauthAccount(
                     $account['client_id'],
@@ -50,10 +49,9 @@ class ApiClientFactory implements RegisteredByNameInterface
         if ($type === 'e5_dal_token') {
             try {
                 // no need to authorize token is fixed
-                $client = new ApiClient($account['domain']);
-                $cachePath = $_ENV['CACHE_PATH'] ?? sys_get_temp_dir();
-                $this->memoryContext = $this->memoryContext ?? new MemoryContext($cachePath.'/e5_connection_memory_context.json');
-                $account = new E5DalAPIAccount($account['token'], $this->memoryContext);
+                $client = new ApiCurlClient($domain);
+
+                $account = new E5DalAPIAccount($account['token']);
                 $client->authorize($account);
 
                 return $client;
@@ -63,7 +61,7 @@ class ApiClientFactory implements RegisteredByNameInterface
         }
 
         try {
-            $client = new ApiClient($account['domain']);
+            $client = new ApiCurlClient($domain);
 
             $account = new AkeneoApiClientAccount(
                 $account['username'],
