@@ -77,26 +77,26 @@ interface ApiClientInterface
     /**
      * A GET HTTP VERB
      */
-    public function get(string $endpoint): self;
+    public function get(string $endpoint): ApiResponse;
     /**
      * A POST HTTP VERB
      * $postData is a structured entity array that will be encoded to json
      */
-    public function post(string $endpoint, array $postData): self;
+    public function post(string $endpoint, array $postData, array $headers = []): ApiResponse;
     /**
      * HTTP PATCH VERB That supports a multi patch insert
      * max 100 inserts per request
      */
-    public function multiPatch(string $endpoint, array $dataSet): self;
+    public function multiPatch(string $endpoint, array $dataSet): ApiResponse;
     /**
      * HTTP PATCH VERB
      */
-    public function patch(string $endpoint, array $patchData): self;
+    public function patch(string $endpoint, array $patchData): ApiResponse;
     /**
      * A DELETE HTTP VERB
      */
-    public function delete(string $endpoint): self;
-    public function getResponse(): ApiResponse;
+    public function delete(string $endpoint): ApiResponse;
+
     public function log(string $message, int $statusCode = null, $content): void;
 
     public function getUrlGenerator(): UrlGenerator;
