@@ -24,9 +24,19 @@ class ArrayCollection
         $this->items[$key] = $item;
     }
 
+    public function containsKey($key): bool
+    {
+        return isset($this->items[$key]);
+    }
+
     public function get($key): self
     {
         return new self([$this->items[$key] ?? null]);
+    }
+
+    public function remove($key): void
+    {
+        unset($this->items[$key]);
     }
 
     public function first()
@@ -34,7 +44,7 @@ class ArrayCollection
         return current($this->items);
     }
 
-    public function merge(ArrayCollection $collection)
+    public function merge(ArrayCollection $collection): void
     {
         foreach ($collection->getValues() as $item) {
             $this->items[] = $item;
@@ -66,6 +76,11 @@ class ArrayCollection
         return $this->items;
     }
 
+    public function toArray(): array
+    {
+        return $this->items;
+    }
+
     public function addValues(array $items): void
     {
         foreach (array_filter($items) as $key => $item) {
@@ -76,5 +91,6 @@ class ArrayCollection
     public function purge(): void
     {
         $this->items = [];
+
     }
 }
