@@ -2,15 +2,16 @@
 
 namespace Misery\Component\Configurator;
 
+#use App\Component\Common\Resource\NamedResourceInterface;
+#use App\Component\Common\Resource\ResourceCollectionInterface;
 use App\Component\ChangeManager\ChangeManager;
 use Misery\Component\Action\ItemActionProcessorFactory;
+use Misery\Component\Common\Client\ApiClientInterface;
 use Misery\Component\Logger\ItemLoggerAwareTrait;
 use Psr\Log\LoggerAwareTrait;
 use Psr\Log\LoggerInterface;
 use Misery\Component\Action\ItemActionProcessor;
 use Misery\Component\BluePrint\BluePrint;
-use Misery\Component\Common\Client\ApiClient;
-use Misery\Component\Common\Client\ApiClientInterface;
 use Misery\Component\Common\Collection\ArrayCollection;
 use Misery\Component\Common\FileManager\LocalFileManager;
 use Misery\Component\Common\Pipeline\Pipeline;
@@ -46,10 +47,11 @@ class Configuration
     private $sources;
     private $context = [];
     private $shellCommands;
-    /** @var ApiClient[] */
-    private $accounts = [];
-    private $isMultiStep = false;
     public ChangeManager $changeManager;
+    /** @var ApiClientInterface[] */
+    private array $accounts = [];
+    private bool $isMultiStep = false;
+    private array $resourceCollections = [];
     public ItemActionProcessorFactory $actionFactory;
 
     private array $extensions = [];
@@ -103,9 +105,14 @@ class Configuration
         $this->isMultiStep = true;
     }
 
-    public function addContext(array $context)
+    public function addContext(array $context): void
     {
         $this->context = array_merge($this->context, $context);
+    }
+
+    public function setContext(string $key, $value): void
+    {
+        $this->context[$key] = $value;
     }
 
     public function getContext(string $key = null)
@@ -123,12 +130,27 @@ class Configuration
         return $this->sources;
     }
 
+//    public function addResourceCollection(ResourceCollectionInterface|NamedResourceInterface $resourceCollection): void
+//    {
+//        $this->resourceCollections[$resourceCollection->getName()] = $resourceCollection;
+//    }
+//
+//    public function getResourceCollection(string $name):? ResourceCollectionInterface
+//    {
+//        return $this->resourceCollections[$name] ?? null;
+//    }
+
+    public function getResourceCollections(): array
+    {
+        return $this->resourceCollections;
+    }
+  
     public function getAccounts(): array
     {
         return $this->accounts;
     }
 
-    public function addAccount(string $name, ApiClientInterface $apiClient)
+    public function addAccount(string $name, ApiClientInterface $apiClient): void
     {
         $this->accounts[$name] = $apiClient;
     }
