@@ -6,13 +6,12 @@ use Misery\Component\Common\Client\ApiClientAccountInterface;
 use Misery\Component\Common\Client\ApiClientInterface;
 use Misery\Component\Common\Client\ApiEndPointsInterface;
 use Misery\Component\Common\Client\AuthenticatedAccount;
-use Misery\Component\Common\Client\MemoryContext;
 use Misery\Component\Common\Client\PaginationCursor;
 use Misery\Component\Connections\E5Dal\Client\Endpoints\E5DalApiEndpoints;
 
 class E5DalAPIAccount implements ApiClientAccountInterface
 {
-    public function __construct(private string $token, private readonly MemoryContext $memoryContext) {}
+    public function __construct(private string $token) {}
 
     public function getToken(): string
     {
@@ -35,7 +34,7 @@ class E5DalAPIAccount implements ApiClientAccountInterface
 
     public function getPaginator(ApiClientInterface $client, string $endpoint): PaginationCursor
     {
-        return new E5PaginationCursor($client, $endpoint, $this->memoryContext);
+        return new E5PaginationCursor($client, $endpoint);
     }
 
     public function refresh(ApiClientInterface $client, AuthenticatedAccount $account): AuthenticatedAccount
