@@ -75,20 +75,10 @@ class ItemParserFactory implements RegisteredByNameInterface
         if ($type === 'xml') {
             return XmlParser::create(
                 $manager->getFile($configuration['filename']),
-                $configuration['container'] ?? null
+                $configuration['xpath'] ?? null
             );
         }
         if ($type === 'csv') {
-            if (isset($configuration['encoding']) && $configuration['encoding'] === 'UTF8-BOM') {
-                return CsvBomParser::create(
-                    $manager->getFile($configuration['filename']),
-                    $configuration['delimiter'] ?? CsvParser::DELIMITER,
-                    $configuration['enclosure'] ?? CsvParser::ENCLOSURE,
-                    $configuration['escape'] ?? CsvParser::ESCAPE,
-                    $configuration['invalid_lines'] ?? CsvParser::INVALID_STOP
-                );
-            }
-
             return CsvParser::create(
                 $manager->getFile($configuration['filename']),
                 $configuration['delimiter'] ?? CsvParser::DELIMITER,
